Add actions returning metrics for services, and a given service.
The ServiceMetricsPresenter is responsible for combining a service with
a time period, which in turn delegates responsibility for fetching, and
aggregating the metrics for that service and time period to the
AggregatedTransactionsReceivedMetric and
AggregatedTransactionsWithOutcomeMetric.

These are all serialized with their respective serializers.

As of this commit the services#index action issues N+1 queries when
fetching the metrics for each service. This could be optimised in the
future to only issue a single query, letting the database do the
aggregation.