Expose department's metrics through the API.
Adds functionality which allows the aggregated metrics classes to
calculate for departments, agencies or individual services. It does this
by using a common interface, `#services`, which returns the services for
the given organisation. In the case of an individual service, it just
returns a scope which resolves to itself.

Currently the DepartmentMetricsPresenter contains a lot of shared code
from ServiceMetricsPresenter, and would be a good candidate for further
refactoring.