Add specs which exercise the API client.
These specs are closer to integration tests than unit tests as they
exercise the whole client, from making a HTTP request to parsing the
response, and building the model objects.

The HTTP responses are stubbed using VCR. In the future we should
provide a mechanism to run the tests without the responses being stubbed
and instead hitting the API directly, as well as a way to refresh the
fixture data.