Rename Agency to DeliveryOrganisation.
Unfortunatly services don't always neatly fit into a strict Department >
Agency > Service hierarchy. Therefore we're replacing the concept of an
agency with a delivery organisation.

A delivery organisation is the organisation tasked with providing the
service. They could be a department, an agency, or another government
body. Usually they'll belong to a department, but in some cases they'll
be the depeartment, or won't belong to a department.

Currently the data model doesn't address this, and still assumes that an
service belongs to a department, and optionally a delivery organisation.
This'll be addressed when tackling importing this hierarchy.