Generalize the metrics API.
The API, and the code that supports it, is now more general and shared
as much as possible.

The API is now access in a more consistent manner. For example:
 * `/government/metrics`
 * `/departments/:id/metrics`
 * `/delivery_organisations/:id/metrics`
 * `/services/:id/metrics`

In addition there's much less duplication in the code, with most of the
logic being moved out of "presenter" classes into "metrics" classes
which now share much more code, being customised using a few overriden
methods in sub-classes.

The controllers are much more simplified, but duplicated, rather than
genericized as there too simple at the moment to warrent the mental
overhead.