Add collapsing metric group behaviour to metric groups.
This simply adds a class to a metric group if it's collapsed. It uses
Javascript (instead of just rendering HTML) as this will in the future
have behaviour which should degrade in the absense of Javascript.

If we rendered it server-side instead then we wouldn't have all of the
necessary HTML, if Javascript wasn't available.