Better looking search bar
Emulating the same search bar we're using on the Digital Marketplace,
which is emulating the search bar on the GOV.UK categories pages
(the one in the header).

Left in all the SCSS variables as comments because they should
start working once we sort out the frontend assets imports.

Looks good, but doesn't do anything yet.