Add javascript for search bar.
Filters down search results on metrics listing page, using
exact string substring matching.
It's not a very robust search and there's no JS fallback
for it (ie, search bar only shows up if javascript is
enabled), but it's super quick feedback and for simple
use cases it should work pretty well.
There's not a whole lot of service data we're using at the
minute so we don't have anything other than simple use cases.

Also moved the script to the bottom of the page for a tiny
efficiency gain.