Add the selected metric item into the collapsed metric group view.
This re-orders and adds comments to the `collapsedMetricGroup`
Javascript to (attempt to) clarify what it's doing.

The metric items HTML description is stored in data attributes and then
copied to the view, using Javascript, when displayed. The rationale for
this is that if Javascript is not available in the browser, then we
don't want to limit the full information in the metric group from being
displayed. So by implementing this functionality in Javascript (rather
than on the server), if Javascript isn't available then it'll gracefully
degrade.