Disable assets debug mode.
In debug mode, Rails will output multiple script tags for each file
included using Sprockets' dependency management. When we use `async`
this causes errors as the order in which they're loaded is no longer
deterministic.

It's more advantageous to have development behave similarly to
production, so instead we just disable debug mode for the asset
pipeline.