Add importer which updates a delivery organisation's department.
Department's are a special case of delivery organisations, they're the parent,
or top-level, organisation. The list of delivery organisations is imported from
the Government Organisation Register, but this lacks the hierarchy information.

In order to create the hierarchy, we supplement this information with our own,
mapping of delivery organisation to department.

This should be imported after the register has been updated, and it takes a CSV
file (on stdin) in the following format:

   Organisation ID,Department ID
   D101,D1
   D30,D1
   D2,D2

Usage:

   cat mapping.csv | rails runner 'DepartmentsImporter.new.import'