Add new submodules to assets path
Files in each of our submodules can now be referenced by starting
with the folder name and then the rest of the file path.

It's worth noting that this also overwrites the list of things
that Rails defaults to precompiling. The current list includes
only explicit dependencies.

The problem we ran into was that one of our new submodules
included an `application.js` file that was automatically trying
to compile, but and throwing an error when it was referencing
some files that don't exist. Explictly naming the files we want
Rails to know about resolves the problem.