Add base CSS file and overwrite assets paths
Fortunately, we can reference the `_core.scss` file in our
product-page-example submodule to build most of our base scss.
All of the imports still work because the submodule directory has
been added to the asset path.

Unfortunately, the image paths are all hardcoded into the original
SCSS files so we need to override them with new paths.