Remove search button
At the moment, the search button functions only as a signifier.
It is telling you that you can use this field to search the
results, but once you start usng the field, everything happens
automatically.
Clicking the button will only rerun the last search, so, from a
practical point of view, it is pretty much pointless.

Now that the label is visible, removing the button shouldn't
impact the affordance of this field.