Add label to search box
This is a prefigurative commit clearing the way for another one
that will further rationalise the layout.

Printing a different input label based on the level of heirarchy,
which we have access to in the presenter's `group_by` parameter.
Added a new method to the presenter that returns a printable name
for the current `group_by`.

Seemed like a change that was easy to isolate.