Include metrics for the given entity in the metrics API response.
Previously it was possible to fetch this data by passing in the
"group_by" parameter which was equivalent to the type of entity the
metrics were being requested for.

Now we include the metrics for the entity, and then all of the "metric
groups", specified by the "group_by" parameter.

If the entity type and the group by parameters are of equivalent types
then the data returned in the `metrics` and `metric_groups` will be
identical.