Insert totals metric group depending on the specified order.
When ordering by a metric item, we insert the total metric group, at the
beginning or end of the list depending on whether 'High to Low' or 'Low
to High' was selected.

There's a special case, when name is the selected sort order. We never
want to see the total ordered by name, so always prepend it – so it's
displayed at the top.