Fix tests and remove sorting for name
We want the default sort order to be descending for numbers, but for
name it should default to ascending.

As name is the first shown field, when switching to a metric the
default (ascending for name) is carried across and makes it impossible
to have a default of descending for numbers.

This PR removes sorting on name for now .. defaults to alphabetical