Adds a custom 404 and 500 page
The new errors controller allows us to render out own error pages,
currently only 404 and 500 status codes.

It may be that the 500 page (at least) will be served upstream somehow
as a custom error page for when Rails is fully-dead, but this will
allow us to have a more useful page in the mean time.