Unflatten the transactions_with_outcome metrics table
Rather than have a single row per entry, we have unflattened the table
so that the quantity is specified using the new outcome field, which can
be 'any' or 'intended'.  This will allow us to support fields missing
because they are not applicable, as well as fields with no value.