Aggregate transactions with outcome metrics accounting for not applicable and not provided.
When aggregating metrics we need to retain the semantics as to whether
the metric values were not applicable or not provided. If there are no
metrics (rows) for that period, then we treat it as not applicable. If
there are metrics but the values are NULL, then we treat it as not
provided.

We also need to seralize these metrics whilst maintaining these
semantics. In this case if it's not applicable then it shouldn't be
serialised, if it's not provided then the attribute should be serialized
with a NULL value.