Adds new metric item to Calls Received.
Adds a new metric item, 'perform-transaction' to the calls received
metric to hold data about Calls made to perform a transaction.

Modifies the importer to (for now) just copy the data from
TransactionsReceived's phone item (which will be the default at some
point if the value isn't provided). Once we have changed the input form
for the data we can use that value instead.

Also includes a migration that will iterate through current data and if
there is not already a perform-transaction item for each total then it
will copy the data from the TransactionReceivedMetric#phone for the same
service and time period.