Remove services from collapsed list when no sort value
When sorting by metric item (for instance, transactions received total)
and showing the collapsed services, we previously showed items that had
no value (or description in the collapsed view).  This PR hides those
items from the list, making the sort act as a sort and filter.

This PR does not change the number of things being shown (total in tab
text) and therefore can look a little strange when it suggest 34
services but only 7 are showing.