Adds an owner to each service.
Allows a service to have an owner so that we can (later) more easily
give a magic link to the person who owns a service.

Adds the User to admin, showing the services they own where
appropriate.  As we don't expect users to log in with these
accounts just yet, the passwords default to a long string
when user accounts are created.