Tools for storing credentials in PaaS User Provided Service
As we can't get relevant environment variables onto the instance before
the blue-green setup happens then we need a process where the information
is available by binding a service.  Luckily PaaS supports user provided
services which are basically just a hash store.

The two provided scripts will interact with that service and it will be
bound by its presence in the manifest file.

We will still need to change how we read the config/env settings