Provide baseline acronyms for departments we know about
By default these acronyms were generated by taking the upper case
letters from the department name, apart from specific cases like
DExEU and DFT which were tweaked after the file was generated.

These mappings are now used in the synchronisation of depts and
delivery orgs if an acronym is not currently set.  We should
attempt to keep the CSV up to date should the acronyms be
horrendously wrong or changed.