Adds skeleton pages for department and delivery org details
Adds two pages, one for departments and one for delivery organisations
which will eventually hold charts of the data for each metric item. For
now these pages just show the aggregated data and don't yet contain
information on the department type and brief overview of what the
dept or agency does.

Also includes some simple tests that it's working, that can be expanded
to check the charts when implemented