Adds functional sparklines to each metric.
Adds sparklines to every metric, but requires the the un-aggregated data we
have is re-structured so that we can look up the time period's data by
metric name.  This might be prohibitively expensive there we are rendering
lots of services/data.

Currently isn't laid out very well, will need to re-style as per the latest
designs to fix that but can be used to show whether the renderings are accurage.