Fix the department count when viewing at Government level.
When we are looking at government aggregated data, we show the number of
departments, and originally this was just departments that have a
delivery organisation. This didn't take account any delivery
organisations that had no services, or no services with published data.

This PR changes the determination of department count to take into
account only those departments with delivery organisations with services
with published data.

The delivery organisations count, and the services count were already
correct.

This raises the issue of what we do with a custom time period.  Do we
only show departments with published data __within__ that time period?
Or do we show them all but it'll show up as no data?