[WIP] Add a page to show missing data.
This introduces a new model to hold the missing data, but it is
currently not populated (instead of contains an example of usage
to be copied when fetching the actual data from the database).

It would make sense to wait for some data access so that we can test
this before we merge it.