Creates a skeleton detail page for each entity type
For each entity, this PR creates a skeleton page, with content but not
yet with any data shown, although the data is available to the page.

Currently this relies a lot on a helper to obtain the label for
different metrics (from the en.yml file) which are used elsewhere.  This
is pending a change to the data model where the labels, and a metric
iterator will be made available directly.