Initial layout of charts for depts, deliveryorgs and services.
Displays the charts for the department page, the delivery organisation
page and the services page.  The charts currently have several issues:

* The legend does not match the styles used for lines (one solid blue
and one dashed blue)
* Not Provided is not shown when data is now provided, we don't really
know at the point of use at the moment.

This still requires the percentage change when there is a previous time
period (not all data has this yet).