Allow import of users from CSV
In order to allow services to have 'owners' this PR contains a rake task
that will import users from a CSV file.  The CSV file must have 3
columns, the first containing the user's name, the second their email
address and the final column the name of the Service.  The last column
must (unfortunately) exactly match the entry in the database.

The rake task should be run using:

    rake users_import:load < users.csv

Piping the csv into stdin will allow us to pipe it across a tunnel to a
running instance.