Add a link to magic link admin to generate a mailto
This PR modifies the magic link admin page to include a mailto link.
This link will launch the administrators mail client with the subject,
recipient (where available) and content already filled in.

As it isn't possible to make the content HTML in a mailto, we just provide the links
and rely on the server to linkify if they really want to.