Add new email metric
Adds a new email metric to the transactions received, as it has been
identified as a valid channel.

The migration currently defaults to setting the new applicability field to false (for
existing services) so that it can be turned on for specific services by
an admin.  This will mean new services have it turned off by default but
we can fix that easily after the initial onboarding of email.  Some
existing services are using the other field for email currently, so we
need to identify those and move the data into the correct field (perhaps
in the migration that sets the applicability to true).