Renames outcomes to accepted and adds rejected
This PR is branched off
https://github.com/alphagov/service-performance/pull/133 so that should
be reviewed/merged first.

This changes the "Transactions process with the user's intended outcome"
to "Transactions processed and accepted". It also adds "Transactions
processed and rejected".  The new field is not yet taken into account
for validation in the publishing form, but it is in the admin and
aggregated views (+ charts).

There is a rake task ```rake update:rejected``` which will update the
rejected value for monthly metrics that have processed and accepted set.